+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/chats/ChatsScreenState.kt
@@ -1,0 +1,6 @@
+package net.thechance.mena.core_chat.presentation.screen.chats
+
+data class ChatsScreenState(
+    val isLoading: Boolean = false,
+    val isSynced: Boolean = false
+)
