--- core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/main/ChatMainState.kt
@@ -1,6 +1,0 @@
-package net.thechance.mena.core_chat.presentation.screen.main
-
-data class ChatMainState(
-    val isLoading: Boolean = false,
-    val isSynced: Boolean = false
-)
